+++ hutool-core/src/test/java/cn/hutool/core/util/StrUtilTest.java
@@ -1,6 +1,5 @@
 package cn.hutool.core.util;
 
-import cn.hutool.core.lang.Console;
 import cn.hutool.core.lang.Dict;
 import org.junit.jupiter.api.Test;
 
@@ -319,23 +318,23 @@ public class StrUtilTest {
 	 * 验证普通字符、中文字符以及Unicode代理对字符的反转行为
 	 */
 	@Test
-	public void reverseSpecialCharactersTest() {
+	public void reverseByCodePointSpecialCharactersTest() {
 		//普通情况-英文字符
-		assertEquals("dcba", StrUtil.reverse("abcd"));
+		assertEquals("dcba", StrUtil.reverseByCodePoint("abcd"));
 
 		//普通情况-中文字符
-		assertEquals("界世好你", StrUtil.reverse("你好世界"));
+		assertEquals("界世好你", StrUtil.reverseByCodePoint("你好世界"));
 
 		//保证Unicode字符语义正确，类似emoji、组合字符
 		//A😊B
 		String emojiStr = "A\uD83D\uDE0AB";
-		String reversedEmoji = StrUtil.reverse(emojiStr);
+		String reversedEmoji = StrUtil.reverseByCodePoint(emojiStr);
 		//B😊A
 		assertEquals("B\uD83D\uDE0AA", reversedEmoji);
 
 		//A🇨🇳B
 		String surrogate = "A\uD83C\uDDE8\uD83C\uDDF3B";
-		String reversedSurrogate = StrUtil.reverse(surrogate);
+		String reversedSurrogate = StrUtil.reverseByCodePoint(surrogate);
 		//B🇨🇳A
 		assertNotEquals("B\uD83C\uDDE8\uD83C\uDDF3A", reversedSurrogate);
 	}