+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### 🐣新特性
 * 【core   】      `ColorUtil`颜色名称DARKGOLD、LIGHTGOLD新增蛇形命名匹配（pr#1400@Github）
 * 【core   】      添加`BeanPath`方法中对“*”的处理逻辑（pr#1412@Gitee）
+* 【core   】      `StrUtil`添加`reverseByCodePoint`方法（pr#4187@Github）
 
 ### 🐞Bug修复
 * 【core   】      修复`Calculator.conversion`方法计算包含科学计数法表达式的值时逻辑有误，结果不符合预期（pr#4172@Github）